Adding client to get timer offset from server tile
The transmit and receive timestamps may need to be corrected at the
client if the server is on another tile and the client compares the
timestamps with a local timer.

The timers across tile are synchronous but will be offset by a constant.
This constant differs when booting from flash versus boot from JTAG.